+++ packages/dev/CHANGELOG.md
@@ -1,6 +1,14 @@
 @lifterlms/dev CHANGELOG
 ========================
 
+UNRELEASED
+----------
+
++ Updated: [**BREAKING**] During prerelease builds the `readme` command will now exit with code `0` and output a warning message instead of previous behavior: exit code `1` with an error message. 
++ Added: The `readme` command now merges additional merge codes as derived from the `parseMainFileMetadata()` utility function.
++ Added: A New command `meta` has been added.
+
+
 v0.0.5 - 2022-05-23
 -------------------